Bring *_OPS subclasses to main package level.
Surfaces the classes `PondingLoadCell2d_OPS` and `PondingLoadCell3d_OPS`
from the `PondingLoadCell_OPS` module. If openseespy is not installed,
importing PyPonding will raise a UserWarning stating such. The warning
may be suppressed by use of the `warnings.filterwarnings` function or
running Python with the appropriate commandline flags:

    python -W ignore::UserWarning [script or other flags]

IPython does not support these commandline options; instead it must be
launched as a module:

    python -W ignore::UserWarning -m IPython [ipython flags]